Use path() in tests making prove non-dir specific.
Previously used $0 (a.k.a $PROGRAM_NAME) to name the migrations file
for each test, but this would create a new entry in the migrations table
if called from a different directory other than the repository root.

Using `realpath` from `Mojo::File` returns the absolute path of the test
regardless of where prove is executed from.